pyproject.toml
[tool.poetry]
name = "gdsc-demo"
version = "0.1.0"
description = ""
authors = ["gokudaisensei <87324237+gokudaisensei@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "gdsc_demo"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.101.1"}
sqlalchemy = "^2.0.20"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg2 = "^2.9.7"
python-multipart = "^0.0.6"


[tool.poetry.group.test.dependencies]
ipython = "^8.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
